+++ examples/dataset_ilids.py
@@ -12,6 +12,7 @@ import os
 
 import cv2 as cv
 import numpy as np
+
 from pygofpid.pipeline import GOFPID
