+++ examples/video.py
@@ -11,6 +11,7 @@ Perimeter intrusion detection on a single video.
 """
 
 import cv2 as cv
+
 from pygofpid.pipeline import GOFPID
